requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
Pillow
tqdm
matplotlib
opencv-python
streamlit